-Refactor the project structure
