+++ __main__.py
@@ -1,4 +1,4 @@
-from .src import app
+from src import app
 
 if __name__ == "__main__":
     app.run()
