+++ pyproject.toml
@@ -28,7 +28,7 @@ pytest = "^5.4.2"
 autopep8 = "^1.5.2"
 
 [tool.poetry.scripts]
-app = 'src.app:runserver'
+app = 'manage:runserver'
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
